% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difMedicalkey.R
\docType{data}
\name{difMedicalkey}
\alias{difMedicalkey}
\title{Key of Correct Answers for \code{difMedicaltest} Data Set}
\format{A data frame with 20 values representing correct answers to items of \code{difMedicaltest}
data set. For more details see \code{\link{difMedicaltest}}.}
\usage{
data(difMedicalkey)
}
\description{
The \code{difMedicalkey} is a vector of factors representing correct answers of
\code{difMedicaltest} data set.
}
\references{
Vlckova, K. (2014). Test and Item Fairness (Unpublished master's thesis).

Drabinova, A. and Martinkova P. (2016). Detection of Differenctial Item Functioning Based on
Non-Linear Regression, Technical Report, V-1229, \url{http://hdl.handle.net/11104/0259498}.
}
\seealso{
\code{\link{difMedical}}, \code{\link{difMedicaltest}}
}
\author{
Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
adela.drabinova@gmail.com \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Karel Zvara \cr
Faculty of Mathematics and Physics, Charles University \cr
}
\keyword{datasets}
