\name{GMATkey}
\alias{GMATkey}
\docType{data}
\title{
Key of Correct Answers for \code{GMATtest} Data Set
}
\description{The \code{GMATkey} is a vector of factors representing correct answers of generated \code{GMATtest} data set based on Graduate Management Admission Test (GMAT) data set (Kingston et al., 1985).
}
\usage{data(GMATkey)}
\format{
  A data frame with 20 values representing correct answers to items of \code{GMATtest} data set. For more details see \code{\link{GMATtest}}.
}
\author{Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University in Prague \cr
adela.drabinova@gmail.com\cr\cr
Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz\cr\cr
Karel Zvara \cr
Faculty of Mathematics and Physics, Charles University in Prague\cr
}
\references{
Kingston, N., Leary, L., and Wightman, L. (1985). An Exploratory Study of the Applicability of Item Response Theory Methods to the Graduate Management Admission Test. ETS Research Report Series, 1985(2) : 1-64.
}
\seealso{
\code{\link{GMATtest}}, \code{\link{GMAT}}
}
\examples{
data(GMATkey)
}
\keyword{datasets}
