% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difNLR.R
\name{plot.difNLR}
\alias{plot.difNLR}
\title{ICC and test statistics plots for difNLR object}
\usage{
\method{plot}{difNLR}(x, plot.type = "cc", item = "all",
  col = c("dodgerblue2", "goldenrod2"), shape = 21, size = 0.8,
  linetype = c(2, 1), title, ...)
}
\arguments{
\item{x}{an object of 'difNLR' class}

\item{plot.type}{character: type of plot to be plotted (either \code{"cc"} for characteristic curve
(default), or \code{"stat"} for test statistics). See \strong{Details}.}

\item{item}{either character (\code{"all"}), or numeric vector, or single number corresponding to column indicators. See \strong{Details}.}

\item{col}{character: single value, or vector of two values representing colors for plot.}

\item{shape}{integer: shape parameter for plot.}

\item{size}{numeric: single number, or vector of two numbers representing line width in plot.}

\item{linetype}{numeric: single number, or vector of two numbers representing line type in plot for reference and focal group.}

\item{title}{string: title of plot.}

\item{...}{other generic parameters for \code{plot()} function.}
}
\description{
Two types of plots are available. The first one is obtained by setting \code{plot.type = "cc"}
(default). The characteristic curve for item specified in \code{item} option is plotted. For default
option \code{"all"} of item, characteristic curves of all converged items are plotted. The drawn
curves represent best model.

The second plot is obtained by setting \code{plot.type = "stat"}. The  test statistics
(either LR-test, or F-test, depends on argument \code{test}) are displayed on the Y axis,
for each coverged item. The detection threshold is displayed by a horizontal line and items
detected as DIF are printed with the red color. Only parameters \code{size} and \code{title}
are used.
}
