% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difNLR-package.R
\docType{package}
\name{difNLR-package}
\alias{difNLR-package}
\alias{_PACKAGE}
\title{DIF and DDF Detection by Non-Linear Regression Models.}
\description{
The difNLR package containts method for detection of differential item functioning (DIF)
based on non-linear regression. Both uniform and non-uniform DIF effects can be detected when considering
one focal group. The method also allows to test the difference in guessing or inattention parameters
between reference and focal group. DIF detection method is based either on likelihood-ratio test, or on F-test
of a submodel. Package also offers method for DIF detection among ordinal data using adjacent category logit or
cumulative logit models. Moreover, the difNLR package containts method for detection of differential distractor
functioning (DDF) based on multinomial log-linear regression model.
}
\details{
Package: difNLR\cr
Type: Package\cr
Version: 1.3.3\cr
Date: 2020-04-23\cr
Depends: R (>= 3.1)\cr
Imports: CTT, ggplot2 (>= 2.2.1), grDevices, methods, msm, nnet, plyr, reshape2, stats, VGAM\cr
Suggests: ShinyItemAnalysis\cr
License: GPL-3\cr
BugReports: \url{https://github.com/adelahladka/difNLR/issues}\cr
Encoding: UTF-8\cr
}
\note{
This package was supported by grant funded by Czech Science foundation under number GJ15-15856Y.
}
\section{Functions}{

\itemize{
  \item \code{\link{ddfMLR}}
  \item \code{\link{difNLR}}
  \item \code{\link{difORD}}
  \item \code{\link{estimNLR}}
  \item \code{\link{formulaNLR}}
  \item \code{\link{MLR}}
  \item \code{\link{NLR}}
  \item \code{\link{ORD}}
  \item \code{\link{startNLR}}
}
}

\section{Datasets}{

\itemize{
  \item \code{\link{GMAT}}
  \item \code{\link{GMAT2}}
  \item \code{\link{MSATB}}
}
}

\references{
Agresti, A. (2010). Analysis of ordinal categorical data. Second edition. John Wiley & Sons.

Drabinova, A. & Martinkova, P. (2017). Detection of Differential Item Functioning with NonLinear Regression:
Non-IRT Approach Accounting for Guessing. Journal of Educational Measurement, 54(4), 498-517,
\url{https://doi.org/10.1111/jedm.12158}.

Hladka, A. & Martinkova, P. (2020). difNLR: Generalized logistic regression models for DIF and DDF detection.
The R journal. Accepted.

Kingston, N., Leary, L., & Wightman, L. (1985). An Exploratory Study of the Applicability of Item Response Theory
Methods to the Graduate Management Admission Test. ETS Research Report Series, 1985(2) : 1-64.

Martinkova, P., Drabinova, A., Liaw, Y. L., Sanders, E. A., McFarland, J. L., & Price, R. M. (2017).
Checking equity: Why Differential Item Functioning Analysis Should Be a Routine Part of Developing Conceptual
Assessments. CBE-Life Sciences Education, 16(2), \url{https://doi.org/10.1187/cbe.16-10-0307}.

Swaminathan, H. & Rogers, H. J. (1990). Detecting Differential Item Functioning Using Logistic Regression Procedures.
Journal of Educational Measurement, 27, 361-370, \url{https://doi.org/10.1111/j.1745-3984.1990.tb00754.x}

Vlckova, K. (2014). Test and Item Fairness (Unpublished master's thesis).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/adelahladka/difNLR/issues}
}

}
\author{
Adela Hladka (nee Drabinova) \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
\email{hladka@cs.cas.cz} \cr

Patricia Martinkova \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
\email{martinkova@cs.cas.cz} \cr
}
