% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difNLR.R
\name{difNLR}
\alias{difNLR}
\alias{print.difNLR}
\alias{print.difNLR}
\alias{fitted.difNLR}
\alias{coef.difNLR}
\alias{logLik.difNLR}
\alias{AIC.difNLR}
\alias{BIC.difNLR}
\alias{residuals.difNLR}
\alias{resid.difNLR}
\title{Performs DIF detection using Non-Linear Regression method.}
\usage{
difNLR(Data, group, focal.name, model, constraints, type = "both", method = "nls",
match = "zscore", anchor = NULL, purify = FALSE, nrIter = 10, test = "LR", alpha = 0.05,
p.adjust.method = "none", start, initboot = T, nrBo = 20)

\method{print}{difNLR}(x, ...)

\method{fitted}{difNLR}(object, item = "all", ...)

\method{coef}{difNLR}(object, ...)

\method{logLik}{difNLR}(object, item = "all", ...)

\method{AIC}{difNLR}(object, item = "all", ...)

\method{BIC}{difNLR}(object, item = "all", ...)

\method{residuals}{difNLR}(object, item = "all", ...)
}
\arguments{
\item{Data}{numeric: either the scored data matrix only, or the scored data
matrix plus the vector of group. See \strong{Details}.}

\item{group}{numeric or character: either the binary vector of group membership or
the column indicator (in \code{Data}) of group membership. See \strong{Details}.}

\item{focal.name}{numeric or character: indicates the level of \code{group} which corresponds to
focal group}

\item{model}{character: generalized logistic regression model to be fitted. See \strong{Details}.}

\item{constraints}{character: which parameters should be the same for both groups. See \strong{Details}.}

\item{type}{character: type of DIF to be tested. Possible values are \code{"both"} (default), \code{"udif"},
\code{"nudif"}, \code{"all"}, or combination of parameters "a", "b", "c" and "d". See \strong{Details}.}

\item{method}{character: method used to estimate parameters. The options are \code{"nls"} for non-linear least
squares (default) and \code{"likelihood"} for maximum likelihood method.}

\item{match}{character or numeric: specifies matching criterion. Can be either \code{"zscore"} (default, standardized total score),
\code{"score"} (total test score), or numeric vector of the same length as number of observations in "Data". See \strong{Details}.}

\item{anchor}{Either \code{NULL} (default) or a vector of item names or item identifiers specifying which items are
currently considered as anchor (DIF free) items. Argument is ignored if \code{match} is not \code{"zscore"} or \code{"score"}.
See \strong{Details}.}

\item{purify}{logical: should the item purification be applied? (default is \code{FALSE}). See \strong{Details}.}

\item{nrIter}{numeric: the maximal number of iterations in the item purification (default is 10).}

\item{test}{character: test to be performed for DIF detection. Can be either \code{"LR"} (default), or \code{"F"}.
See \strong{Details}.}

\item{alpha}{numeric: significance level (default is 0.05).}

\item{p.adjust.method}{character: method for multiple comparison correction. See \strong{Details}.}

\item{start}{numeric: list with as many elements as number of items. Each element is a named
numeric vector with values representing initial values for parameter estimation. See \strong{Details}.}

\item{initboot}{logical: in case of convergence issues, should be starting values recalculated based on
bootstraped samples? (default is \code{TRUE}). See \strong{Details}.}

\item{nrBo}{numeric: the maximal number of iterations for calculation of starting values using
bootstraped samples (default is 20).}

\item{x}{an object of "difNLR" class}

\item{...}{other generic parameters for S3 methods.}

\item{object}{an object of "difNLR" class}

\item{item}{either character (\code{"all"}), or numeric vector, or single number
corresponding to column indicators.}
}
\value{
The \code{difNLR()} function returns an object of class \code{"difNLR"}. The output
is displayed by the \code{print()} method.

Item characteristic curves and graphical representation of DIF statistics can be displayed
with \code{plot()} method. For more details see \code{\link[difNLR]{plot.difNLR}}.
Estimated parameters can be displayed with \code{coef()} method.

Fitted values can be extracted by the \code{fitted()} method for converged item(s)
specified in \code{item} argument.

Predicted values are produced by the \code{predict()} method for converged item(s)
specified in \code{item} argument. New data can be introduced with \code{match} and \code{group} arguments.
For more details see \code{\link[difNLR]{predict.difNLR}}.

Residuals are extracted with the \code{residuals()} method for converged item(s)
specified in \code{item} argument.

Log-likelihood, Akaike's information criterion and Schwarz's Bayesian criterion can be
extracted with methods \code{logLik()}, \code{AIC()}, \code{BIC()} for converged item(s)
specified in \code{item} argument.

Object of class \code{"difNLR"} is a list with the following components:
\describe{
  \item{\code{Sval}}{the values of test statistics.}
  \item{\code{nlrPAR}}{the estimates of final model.}
  \item{\code{nlrSE}}{the standard errors of estimates of final model.}
  \item{\code{parM0}}{the estimates of null model.}
  \item{\code{seM0}}{the standard errors of estimates of null model.}
  \item{\code{covM0}}{the covariance matrices of estimates of null model.}
  \item{\code{parM1}}{the estimates of alternative model.}
  \item{\code{seM1}}{the standard errors of estimates of alternative model.}
  \item{\code{covM1}}{the covariance matrices of estimates of alternative model.}
  \item{\code{alpha}}{numeric: significance level.}
  \item{\code{DIFitems}}{either the column indicators of the items which were detected as DIF, or
  \code{"No DIF item detected"}.}
  \item{\code{match}}{matching criterion.}
  \item{\code{model}}{fitted model.}
  \item{\code{type}}{character: type of DIF that was tested. If parameters were specified, the value is \code{"other"}.}
  \item{\code{types}}{character: the parameters (specified by user, \code{type} has value \code{"other"}) which were
  tested for difference.}
  \item{\code{p.adjust.method}}{character: method for multiple comparison correction which was applied.}
  \item{\code{pval}}{the p-values by likelihood ratio test.}
  \item{\code{adj.pval}}{the adjusted p-values by likelihood ratio test using \code{p.adjust.method}.}
  \item{\code{df}}{the degress of freedom of likelihood ratio test.}
  \item{\code{test}}{used test.}
  \item{\code{purification}}{\code{purify} value.}
  \item{\code{nrPur}}{number of iterations in item purification process. Returned only if \code{purify}
  is \code{TRUE}.}
  \item{\code{difPur}}{a binary matrix with one row per iteration of item purification and one column per item.
  "1" in i-th row and j-th column means that j-th item was identified as DIF in i-1-th iteration. Returned only
  if \code{purify} is \code{TRUE}.}
  \item{\code{conv.puri}}{logical: indicating whether item purification process converged before the maximal number
  \code{nrIter} of iterations. Returned only if \code{purify} is \code{TRUE}.}
  \item{\code{group}}{the vector of group membership.}
  \item{\code{Data}}{the data matrix.}
  \item{\code{method}}{used estimation method.}
  \item{\code{conv.fail}}{numeric: number of convergence issues.}
  \item{\code{conv.fail.which}}{the indicators of the items which did not converge.}
  \item{\code{llM0}}{log-likelihood of null model.}
  \item{\code{llM1}}{log-likelihood of alternative model.}
}
}
\description{
Performs DIF detection procedure based on non-linear regression model (generalized logistic
regression) and either likelihood-ratio or F test of submodel.
}
\details{
DIF detection procedure based on non-linear regression is the extension of logistic regression
procedure (Swaminathan and Rogers, 1990).

The \code{Data} is a matrix which rows represents scored examinee answers (\code{"1"} correct,
\code{"0"} incorrect) and columns correspond to the items. In addition, \code{Data} can hold
the vector of group membership. If so, \code{group} is a column indicator of \code{Data}.
Otherwise, \code{group} must be a dichotomous vector of the same length as \code{nrow(Data)}.

The unconstrained form of 4PL generalized logistic regression model for probability of correct
answer (i.e., y = 1) is

P(y = 1) = (c + cDif*g) + (d + dDif*g - c - cDif*g)/(1 + exp(-(a + aDif*g)*(x - b - bDif*g))),

where x is by default standardized total score (also called Z-score) and g is group membership.
Parameters a, b, c and d are discrimination, difficulty, guessing and inattention.
Terms aDif, bDif, cDif and dDif then represent differences between two groups in relevant
parameters.

This 4PL model can be further constrained by \code{model} and \code{constraints} arguments.
The arguments \code{model} and \code{constraints} can be also combined.

The \code{model} argument offers several predefined models. The options are as follows:
\code{Rasch} for 1PL model with discrimination parameter fixed on value 1 for both groups,
\code{1PL} for 1PL model with discrimination parameter fixed for both groups,
\code{2PL} for logistic regression model,
\code{3PLcg} for 3PL model with fixed guessing for both groups,
\code{3PLdg} for 3PL model with fixed inattention for both groups,
\code{3PLc} (alternatively also \code{3PL}) for 3PL regression model with guessing parameter,
\code{3PLd} for 3PL model with inattention parameter,
\code{4PLcgdg} for 4PL model with fixed guessing and inattention parameter for both groups,
\code{4PLcgd} (alternatively also \code{4PLd}) for 4PL model with fixed guessing for both groups,
\code{4PLcdg} (alternatively also \code{4PLc}) for 4PL model with fixed inattention for both groups,
or \code{4PL} for 4PL model.

The \code{model} can be specified in more detail with \code{constraints} argument which specifies what
parameters should be fixed for both groups. For example, choice \code{"ad"} means that discrimination (a) and
inattention (d) are fixed for both groups and other parameters (b and c) are not. The arguments
\code{model} and \code{constraints} can be also item specific if they take a form of vector, where
each element correspond to one item. The \code{NA} value for \code{constraints} means no constraints.

The \code{type} corresponds to type of DIF to be tested. Possible values are
\code{"both"} to detect any DIF caused by difference in difficulty or discrimination (i.e., uniform and/or non-uniform),
\code{"udif"} to detect only uniform DIF (i.e., difference in difficulty b),
\code{"nudif"} to detect only non-uniform DIF (i.e., difference in discrimination a), or
\code{"all"} to detect DIF caused by difference caused by any parameter that can differed between groups. The \code{type}
of DIF can be also specified in more detail by using combination of parameters a, b, c and d. For example, with an option
\code{"c"} for 4PL model only the difference in parameter c is tested. The \code{type} argument is
also item specific.

Argument \code{match} represents the matching criterion. It can be either the standardized test score (default, \code{"zscore"}),
total test score (\code{"score"}), or any other continuous or discrete variable of the same length as number of observations
in \code{Data}.

A set of anchor items (DIF free) can be specified through the \code{anchor} argument. It need to be a vector of either
item names (as specified in column names of \code{Data}) or item identifiers (integers specifying the column number).
In case anchor items are provided, only these items are used to compute matching criterion \code{match}. If the \code{match}
argument is not either \code{"zscore"} or \code{"score"}, \code{anchor} argument is ignored. When anchor items are
provided, purification is not applied.

The \code{p.adjust.method} is a character for \code{\link[stats]{p.adjust}} function from the \code{stats}
package. Possible values are \code{"holm"}, \code{"hochberg"}, \code{"hommel"},
\code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"}, \code{"none"}.

The \code{start} is a list with as many elements as number of items. Each element is a named numeric
vector of length 8 representing initial values for parameter estimation. Specifically, parameters
a, b, c, and d are initial values for discrimination, difficulty, guessing and inattention
for reference group. Parameters aDif, bDif, cDif and dDif are then differences in these
parameters between reference and focal group. If not specified, starting
values are calculated with \code{\link[difNLR]{startNLR}} function.

Missing values are allowed but discarded for item estimation. They must be coded as
\code{NA} for both, \code{data} and \code{group} parameters.

In case of convergence issues, with an option \code{initboot = TRUE}, the starting values are
re-calculated based on bootstraped samples. Newly calculated initial values are applied only to
items/models with convergence issues.

In case that model considers difference in guessing or inattention parameter, the different parameterization is
used and parameters with standard errors are recalculated by delta method. However, covariance matrices stick with
alternative parameterization.
}
\examples{
\dontrun{
# loading data based on GMAT
data(GMAT)

Data  <- GMAT[, 1:20]
group <- GMAT[, "group"]

# Testing both DIF effects using likelihood-ratio test and
# 3PL model with fixed guessing for groups
(x <- difNLR(Data, group, focal.name = 1, model = "3PLcg"))

# Testing both DIF effects using F test and
# 3PL model with fixed guessing for groups
difNLR(Data, group, focal.name = 1, model = "3PLcg", test = "F")

# Testing both DIF effects using LR test,
# 3PL model with fixed guessing for groups
# and Benjamini-Hochberg correction
difNLR(Data, group, focal.name = 1, model = "3PLcg", p.adjust.method = "BH")

# Testing both DIF effects using LR test,
# 3PL model with fixed guessing for groups
# and item purification
difNLR(Data, group, focal.name = 1, model = "3PLcg", purify = T)

# Testing both DIF effects using 3PL model with fixed guessing for groups
# and total score as matching criterion
difNLR(Data, group, focal.name = 1, model = "3PLcg", match = "score")

# Testing uniform DIF effects using 4PL model with the same
# guessing and inattention
difNLR(Data, group, focal.name = 1, model = "4PLcgdg", type = "udif")

# Testing non-uniform DIF effects using 2PL model
difNLR(Data, group, focal.name = 1, model = "2PL", type = "nudif")

# Testing difference in parameter b using 4PL model with fixed
# a and c parameters
difNLR(Data, group, focal.name = 1, model = "4PL", constraints = "ac", type = "b")

# Testing both DIF effects using LR test,
# 3PL model with fixed guessing for groups
# with maximum likelihood estimation method
difNLR(Data, group, focal.name = 1, model = "3PLcg", method = "likelihood")

# Graphical devices
plot(x)
plot(x, item = x$DIFitems)
plot(x, plot.type = "stat")

# Coefficients
coef(x)

# Fitted values
fitted(x)
fitted(x, item = 1)

# Residuals
residuals(x)
residuals(x, item = 1)

# Predicted values
predict(x)
predict(x, item = 1)

# Predicted values for new subjects
predict(x, item = 1, match = 0, group = 1)
predict(x, item = 1, match = 0, group = 0)

# AIC, BIC, logLik
AIC(x); AIC(x, item = 1)
BIC(x); BIC(x, item = 1)
logLik(x); logLik(x, item = 1)
}

}
\references{
Drabinova, A. & Martinkova P. (2017). Detection of Differential Item Functioning with NonLinear Regression:
Non-IRT Approach Accounting for Guessing. Journal of Educational Measurement, 54(4), 498-517,
\url{https://doi.org/10.1111/jedm.12158}.

Swaminathan, H. & Rogers, H. J. (1990). Detecting Differential Item Functioning Using Logistic Regression Procedures.
Journal of Educational Measurement, 27, 361-370.
}
\seealso{
\code{\link[stats]{p.adjust}} \code{\link[difNLR]{plot.difNLR}} \code{\link[difNLR]{startNLR}}
}
\author{
Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
drabinova@cs.cas.cz \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Karel Zvara \cr
Faculty of Mathematics and Physics, Charles University \cr
}
\keyword{DIF}
