% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimNLR.R
\name{estimNLR}
\alias{estimNLR}
\title{Estimation for non-linear regression DIF models}
\usage{
estimNLR(y, match, group, formula, start, method)
}
\arguments{
\item{y}{numeric: binary vector. See \strong{Details}.}

\item{match}{numeric: matching criterion. See \strong{Details}.}

\item{group}{numeric: binary vector of group membership. \code{"0"} for reference group, \code{"1"} for focal group.}

\item{formula}{list: with the same structure as output of \code{formulaNLR} function. See \strong{Details}.}

\item{start}{numeric: initial parameters. See \strong{Details}.}

\item{method}{character: method used to estimate parameters. Options are \code{"nls"} for non-linear
least squares and \code{"likelihood"} for maximum likelihood. See \strong{Details}.}
}
\description{
Estimates parameters of non-linear regression models for DIF detection.
This is very first version of the function.
}
\examples{
\dontrun{
data(GMAT)

y  <- GMAT[, 1]
x  <- scale(apply(GMAT[, 1:20], 1, sum))
g <- GMAT[, "group"]

# starting values for 3PL model
start <- startNLR(GMAT[, 1:20], g, model = "3PLcg")
start <- start[1, c(1, 2, 3, 5, 6)]
names(start) <- c("a", "b", "c", "aDif", "bDif")

formula <- formulaNLR(model = "3PLcg", type = "both")

# Non-linear least squares
fitNLS <- estimNLR(y = y, match = x, group = g, formula = formula,
start = start, method = "nls")
lapply(fitNLS, coef)
lapply(fitNLS, logLik)
lapply(fitNLS, vcov)

# Maximum likelihood
fitLKL <- estimNLR(y = y, match = x, group = g, formula = formula,
start = start, method = "likelihood")
lapply(fitLKL, coef)
lapply(fitLKL, logLik)
lapply(fitLKL, vcov)
}

}
\author{
Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
drabinova@cs.cas.cz \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr
}
\keyword{DIF}
