% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMAT.R
\docType{data}
\name{GMAT}
\alias{GMAT}
\title{Dichotomous Data Set Based on Graduate Management Admission Test with the Same Total Score Distribution for Groups}
\format{A \code{GMAT} data frame consists of 2,000 observations on the following 21 variables.
The first 20 columns represent dichotomously scored items of the test. The 21st column is vector
of group membership; values 0 and 1 refer to reference and focal group.}
\usage{
data(GMAT)
}
\description{
The \code{GMAT} data set is generated data set based on parameters from Graduate
Management Admission Test (GMAT) data set (Kingston et al., 1985). First two items were
considered to function differently in uniform and non-uniform way respectively. The data set
represents responses of 2,000 subjects to multiple-choice test of 20 items. A correct answer
is coded as 1 and incorrect answer as 0. The column \code{group} represents group membership,
where 0 represents reference group and 1 represents focal group. Groups are the same
size (i.e. 1,000 per group). The distributions of total scores (sum of correct answers) are the
same for both reference and focal group (Martinkova et al., 2016).
}
\author{
Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
adela.drabinova@gmail.com \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Karel Zvara \cr
Faculty of Mathematics and Physics, Charles University \cr
}
\references{
Kingston, N., Leary, L., and Wightman, L. (1985). An Exploratory Study of the Applicability of Item Response Theory Methods to the Graduate Management Admission Test. ETS Research Report Series, 1985(2) : 1-64.

Martinkova, P., Drabinova, A., Liaw Y.-L., Sanders E. A., McFarland J. L., Price R. M. (2016). Using DIF Analysis to Reveal Potential Equity Gaps in Conceptual Assessments. In review.
}
\seealso{
\code{\link{GMATtest}}, \code{\link{GMATkey}}
}
\keyword{datasets}

