% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMAT2key.R
\docType{data}
\name{GMAT2key}
\alias{GMAT2key}
\title{Key of Correct Answers for \code{GMAT2test} Data Set}
\format{A data frame with 20 values representing correct answers to items of \code{GMAT2test} data set.
For more details see \code{\link{GMAT2test}}.}
\usage{
data(GMAT2key)
}
\description{
The \code{GMAT2key} is a vector of factors representing correct answers of generated
\code{GMAT2test} data set based on Graduate Management Admission Test (GMAT) data set (Kingston et al., 1985).
}
\references{
Kingston, N., Leary, L., and Wightman, L. (1985). An Exploratory Study of the Applicability of Item Response
Theory Methods to the Graduate Management Admission Test. ETS Research Report Series, 1985(2) : 1-64.

Drabinova, A. and Martinkova P. (2016). Detection of Differenctial Item Functioning Based on Non-Linear
Regression, Technical Report, V-1229, \url{http://hdl.handle.net/11104/0259498}.
}
\seealso{
\code{\link{GMAT2}}, \code{\link{GMAT2test}}
}
\author{
Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
adela.drabinova@gmail.com \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr
}
\keyword{datasets}
