% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMAT2key.R
\docType{data}
\name{GMAT2key}
\alias{GMAT2key}
\title{Key of correct answers for \code{GMAT2test} dataset.}
\format{A nominal vector with 20 values representing correct answers to items of \code{GMAT2test}
dataset. For more details see \code{\link{GMAT2test}}.}
\usage{
data(GMAT2key)
}
\description{
The \code{GMAT2key} is a vector of factors representing correct answers of generated
\code{GMAT2test} data set based on Graduate Management Admission Test (GMAT) data set (Kingston et al., 1985).
}
\references{
Kingston, N., Leary, L., & Wightman, L. (1985). An Exploratory Study of the Applicability of Item Response
Theory Methods to the Graduate Management Admission Test. ETS Research Report Series, 1985(2) : 1-64.

Drabinova, A. & Martinkova P. (2017). Detection of Differential Item Functioning with NonLinear Regression:
Non-IRT Approach Accounting for Guessing. Journal of Educational Measurement, 54(4), 498-517,
\url{https://doi.org/10.1111/jedm.12158}.
}
\seealso{
\code{\link{GMAT2}}, \code{\link{GMAT2test}}
}
\author{
Adela Hladka (nee Drabinova) \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
\email{hladka@cs.cas.cz} \cr

Patricia Martinkova \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
\email{martinkova@cs.cas.cz} \cr
}
\keyword{datasets}
