% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMATkey.R
\docType{data}
\name{GMATkey}
\alias{GMATkey}
\title{Key of Correct Answers for \code{GMATtest} Data Set}
\format{A data frame with 20 values representing correct answers to items of \code{GMATtest}
data set. For more details see \code{\link{GMATtest}}.}
\usage{
data(GMATkey)
}
\description{
The \code{GMATkey} is a vector of factors representing correct answers of
generated \code{GMATtest} data set based on Graduate Management Admission Test (GMAT)
data set (Kingston et al., 1985).
}
\author{
Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
adela.drabinova@gmail.com \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Karel Zvara \cr
Faculty of Mathematics and Physics, Charles University \cr
}
\references{
Kingston, N., Leary, L., and Wightman, L. (1985). An Exploratory Study of the Applicability of Item Response Theory Methods to the Graduate Management Admission Test. ETS Research Report Series, 1985(2) : 1-64.

Martinkova, P., Drabinova, A., Liaw Y.-L., Sanders E. A., McFarland J. L., Price R. M. (2016). Using DIF Analysis to Reveal Potential Equity Gaps in Conceptual Assessments. In review.
}
\seealso{
\code{\link{GMAT}}, \code{\link{GMATtest}}
}
\keyword{datasets}

