% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMATkey.R
\docType{data}
\name{GMATkey}
\alias{GMATkey}
\title{Key of correct answers for \code{GMATtest} dataset.}
\format{A nominal vector with 20 values representing correct answers to items of \code{GMATtest}
dataset. For more details see \code{\link{GMATtest}}.}
\usage{
data(GMATkey)
}
\description{
The \code{GMATkey} is a vector of factors representing correct answers of
generated \code{GMATtest} data set based on Graduate Management Admission Test (GMAT, Kingston et al., 1985).
}
\references{
Kingston, N., Leary, L., & Wightman, L. (1985). An Exploratory Study of the Applicability of Item Response
Theory Methods to the Graduate Management Admission Test. ETS Research Report Series, 1985(2) : 1-64.

Martinkova, P., Drabinova, A., Liaw, Y. L., Sanders, E. A., McFarland, J. L., & Price, R. M. (2017).
Checking equity: Why Differential Item Functioning Analysis Should Be a Routine Part of Developing Conceptual
Assessments. CBE-Life Sciences Education, 16(2), rm2, \url{https://doi.org/10.1187/cbe.16-10-0307}.
}
\seealso{
\code{\link{GMAT}}, \code{\link{GMATtest}}
}
\author{
Adela Hladka (nee Drabinova) \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
\email{hladka@cs.cas.cz} \cr

Patricia Martinkova \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
\email{martinkova@cs.cas.cz} \cr
}
\keyword{datasets}
