% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSATBkey.R
\docType{data}
\name{MSATBkey}
\alias{MSATBkey}
\title{Key of correct answers for \code{MSATBtest} dataset.}
\format{A nominal vector with 20 values representing correct answers to items of \code{MSATBtest}
dataset. For more details see \code{\link{MSATBtest}}.}
\usage{
data(MSATBkey)
}
\description{
The \code{MSATBkey} is a vector of factors representing correct answers of
\code{MSATBtest} dataset.
}
\references{
Drabinova, A. & Martinkova P. (2017). Detection of Differential Item Functioning with NonLinear Regression:
Non-IRT Approach Accounting for Guessing. Journal of Educational Measurement, 54(4), 498-517,
\url{https://doi.org/10.1111/jedm.12158}.

Vlckova, K. (2014). Test and Item Fairness (Unpublished master's thesis).
}
\seealso{
\code{\link{MSATB}}, \code{\link{MSATBtest}}
}
\author{
Adela Hladka (nee Drabinova) \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
\email{hladka@cs.cas.cz} \cr

Patricia Martinkova \cr
Institute of Computer Science of the Czech Academy of Sciences \cr
\email{martinkova@cs.cas.cz} \cr
}
\keyword{datasets}
