% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLR.R
\name{NLR}
\alias{NLR}
\title{Non-Linear Regression DIF statistic}
\usage{
NLR(Data, group, model, type = "both", start,
p.adjust.method = "none", test = "LR", alpha = 0.05)
}
\arguments{
\item{Data}{numeric: either binary data matrix only, or the binary data matrix plus the vector of group . See \strong{Details}.}

\item{group}{numeric: binary vector of group membership. "0" for reference group, "1" for focal group.}

\item{model}{character: generalized logistic regression model to be fitted. See \strong{Details}.}

\item{type}{character: type of DIF to be tested (either "both" (default), "udif", or "nudif"). See \strong{Details}.}

\item{start}{numeric: matrix with n rows (where n is the number of items) and at most 5 columns containing initial item parameters estimates. See \strong{Details}.}

\item{p.adjust.method}{character: method for multiple comparison correction. See \strong{Details}.}

\item{test}{character: test to be performed for DIF detection (either "F" (default), or "LR"). See \strong{Details}.}

\item{alpha}{numeric: significance level (default is 0.05).}
}
\description{
Performs DIF detection procedure based on Non-Linear Regression and either F-test or likelihood ratio test of submodel.
}
\details{
DIF detection procedure based on Non-Linear Regression is the extension
of Logistic Regression procedure (Swaminathan and Rogers, 1990).

The \code{Data} is a matrix whose rows represents examinee scored answers
("1" correct, "0" incorrect) and columns correspond to the items.
The \code{group} must be a vector of the same length as \code{nrow(data)}.

The options of \code{model} are as follows: \code{Rasch} for one-parameter logistic model with
discrimination parameter fixed on value 1 for both groups, \code{1PL} for one-parameter logistic
model with discrimination parameter fixed for both groups, \code{2PL} for logistic regression model,
\code{3PLcg} for three-parameter logistic regression model with fixed guessing for both groups,
\code{3PLdg} for three-parameter logistic regression model with fixed inattention for both groups, or
\code{4PLcgdg} for four-parameter logistic regression model with fixed guessing and inattention
parameter for both groups.
The \code{type} corresponds to type of DIF to be tested. Possible values
are \code{"both"} to detect any DIF (uniform and/or non-uniform), \code{"udif"}
to detect only uniform DIF or \code{"nudif"} to detect only non-uniform DIF.

The \code{start} is a matrix with a number of rows equal to number of items and with 8 columns.
First 4 columns represent parameters (a, b, c, d) of generalized logistic regression model
for reference group. Last 4 columns represent differences of parameters (aDif, bDif, cDif, dDif)
of generalized logistic regression model between reference and focal group. If \code{start}
missing, starting values are calculated by \code{startNLR()} function.

The \code{p.adjust.method} is a character for \code{p.adjust} function from the
\code{stats} package. Possible values are \code{"holm"}, \code{"hochberg"},
\code{"hommel"}, \code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"},
\code{"none"}.

Missing values are allowed but discarded for item estimation. They must be coded
as \code{NA} for both, \code{data} and \code{group} parameters.
}
\examples{
\dontrun{
# loading data based on GMAT
data(GMAT)

data  <- GMAT[, colnames(GMAT) != "group"]
group <- GMAT[, "group"]

# Testing both DIF effects using LR test (default)
# and model with fixed guessing for both groups
NLR(Data, group, model = "3PLcg")

# using F test
NLR(Data, group, model = "3PLcg", test = "F")

# Testing both DIF effects with Benjamini-Hochberg correction
NLR(Data, group, model = "3PLcg", p.adjust.method = "BH")

# Testing uniform DIF effects
NLR(Data, group, model = "3PLcg", type = "udif")
# Testing non-uniform DIF effects
NLR(Data, group, model = "3PLcg", type = "nudif")
}
}
\author{
Adela Drabinova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
adela.drabinova@gmail.com \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr

Karel Zvara \cr
Faculty of Mathematics and Physics, Charles University \cr
}
\references{
Drabinova, A. and Martinkova P. (2016). Detection of Differenctial Item Functioning Based on Non-Linear Regression, Technical Report, V-1229, \url{http://hdl.handle.net/11104/0259498}.

Swaminathan, H. and Rogers, H. J. (1990). Detecting Differential Item Functioning Using Logistic Regression Procedures. Journal of Educational Measurement, 27, 361-370.
}
\seealso{
\code{\link[stats]{p.adjust}}
}
\keyword{DIF}

