% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORD.R
\name{ORD}
\alias{ORD}
\title{DDF likelihood ratio statistics for ordinal data.}
\usage{
ORD(Data, group, model = "adjacent", type = "both", match = "zscore",
anchor = 1:ncol(Data), p.adjust.method = "none", parametrization = "irt",
alpha = 0.05)
}
\arguments{
\item{Data}{matrix or data.frame: the ordinarily scored data matrix.}

\item{group}{numeric or character: the binary vector of group membership}

\item{model}{character: logistic regression model for ordinal data (either \code{"adjacent"} (default) or \code{"cumulative"}).
See \strong{Details}.}

\item{type}{character: type of DDF to be tested (either \code{"both"} (default), \code{"udif"}, or \code{"nudif"}).
See \strong{Details}.}

\item{match}{specifies matching criterion. Can be either \code{"zscore"} (default, standardized total score),
\code{"score"} (total test score), or vector of the same length as number of observations in \code{Data}. See \strong{Details}.}

\item{anchor}{a vector of integers specifying which items are currently considered as anchor (DDF free) items. By
default, all items are considered as anchors. Argument is ignored if \code{match} is not \code{"zscore"} or \code{"score"}.
See \strong{Details}.}

\item{p.adjust.method}{character: method for multiple comparison correction.
See \strong{Details}.}

\item{parametrization}{character: parametrization of regression coefficients. Possible options are
\code{"irt"} (default) and \code{"classic"}. See \strong{Details}.}

\item{alpha}{numeric: significance level (default is 0.05).}
}
\value{
A list with the following arguments:
\describe{
  \item{\code{Sval}}{the values of likelihood ratio test statistics.}
  \item{\code{pval}}{the p-values by likelihood ratio test.}
  \item{\code{adj.pval}}{the adjusted p-values by likelihood ratio test using \code{p.adjust.method}.}
  \item{\code{df}}{the degress of freedom of likelihood ratio test.}
  \item{\code{par.m0}}{the estimates of null model.}
  \item{\code{par.m1}}{the estimates of alternative model.}
  \item{\code{se.m0}}{standard errors of parameters in null model.}
  \item{\code{se.m1}}{standard errors of parameters in alternative model.}
  \item{\code{ll.m0}}{log-likelihood of null model.}
  \item{\code{ll.m1}}{log-likelihood of alternative model.}
  \item{\code{AIC.m0}}{AIC of null model.}
  \item{\code{AIC.m1}}{AIC of alternative model.}
  \item{\code{BIC.m0}}{BIC of null model.}
  \item{\code{BIC.m1}}{BIC of alternative model.}
  }
}
\description{
Calculates DDF likelihood ratio statistics for ordinal data based either on adjacent logistic regression
model or on cumulative logistic regression model.
}
\details{
Calculates DDF likelihood ratio statistics for ordinal data based either on adjacent logistic regression
model or on cumulative logistic regression model

The \code{Data} is a matrix or data.frame which rows represents examinee ordinarily scored answers and
columns correspond to the items. The \code{group} must be a vector of the same length as \code{nrow(Data)}.

The \code{model} corresponds to model to be used for DDF detection. Options are \code{"adjacent"}
for adjacent logistic regression model or \code{"cumulative"} for cumulative logistic regression model.

The \code{type} corresponds to type of DDF to be tested. Possible values are \code{"both"}
to detect any DDF (uniform and/or non-uniform), \code{"udif"} to detect only uniform DDF or
\code{"nudif"} to detect only non-uniform DDF.

Argument \code{match} represents the matching criterion. It can be either the standardized test score (default, \code{"zscore"}),
total test score (\code{"score"}), or any other continuous or discrete variable of the same length as number of observations
in \code{Data}.

The \code{p.adjust.method} is a character for \code{p.adjust} function from the \code{stats} package. Possible values are
\code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"}, and
\code{"none"}. See also \code{\link[stats]{p.adjust}} for more information.

Argument \code{parametrization} is a character which specifies parametrization of regression parameters. Default option
is \code{"irt"} which returns IRT parametrization (difficulty-discrimination). Option \code{"classic"} returns
intercept-slope parametrization with effect of group membership and interaction with matching criterion.
}
\examples{
\dontrun{
# loading data
data(dataMedicalgraded, package = "ShinyItemAnalysis")
df <- dataMedicalgraded[, c(1:5, 101)]
df <- df[complete.cases(df), ]

Data <- df[, 1:5]
group <- df[, 6]

# Testing both DDF effects
ORD(Data, group, type = "both")

# Testing uniform DDF effects
ORD(Data, group, type = "udif")

# Testing non-uniform DDF effects
ORD(Data, group, type = "nudif")

# Testing DDF using cumulative logit model
ORD(Data, group, model = "cumulative")
}

}
\seealso{
\code{\link[stats]{p.adjust}} \code{\link[VGAM]{vglm}}
}
\author{
Adela Hladka (nee Drabinova) \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
Faculty of Mathematics and Physics, Charles University \cr
hladka@cs.cas.cz \cr

Patricia Martinkova \cr
Institute of Computer Science, The Czech Academy of Sciences \cr
martinkova@cs.cas.cz \cr
}
\keyword{DDF}
